requirements.txt
pvlib==0.10.3
requests==2.31.0
paho-mqtt==1.6.1
influxdb-client==3.0.0
schedule==1.2.1
pytz==2024.1
pandas==2.2.0
numpy==1.26.4
google-generativeai==0.4.1
pydantic==2.6.1
PyYAML==6.0.1
SQLAlchemy==2.0.27
aiohttp==3.9.3
async-timeout==4.0.3
dataclasses-json==0.6.4
jsonpatch==1.33
tenacity==8.2.3
typing-extensions==4.9.0
jsonpointer==2.4
marshmallow==3.20.2
mypy-extensions==1.0.0
packaging==23.2
sniffio==1.3.0
idna==3.6
multidict==6.0.5
yarl==1.9.4
frozenlist==1.4.1
aiosignal==1.3.1
attrs==23.2.0
